// src/frontend/src/app/api/fn/submissions/api-exams-exam-id-submissions-post-json.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { SubmissionDto } from '../../models/submission-dto';

export interface ApiExamsExamIdSubmissionsPost$Json$Params {
  examId: number;
      body?: {
'StudentName'?: string;
'StudentCode'?: string;
'File'?: Blob;
}
}

export function apiExamsExamIdSubmissionsPost$Json(http: HttpClient, rootUrl: string, params: ApiExamsExamIdSubmissionsPost$Json$Params, context?: HttpContext): Observable<StrictHttpResponse<SubmissionDto>> {
  const rb = new RequestBuilder(rootUrl, apiExamsExamIdSubmissionsPost$Json.PATH, 'post');
  if (params) {
    rb.path('examId', params.examId, {});
    rb.body(params.body, 'multipart/form-data');
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'text/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<SubmissionDto>;
    })
  );
}

apiExamsExamIdSubmissionsPost$Json.PATH = '/api/exams/{examId}/submissions';