// src/frontend/src/app/api/fn/submissions/api-submissions-id-patch-json.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { SubmissionDto } from '../../models/submission-dto';

export interface ApiSubmissionsIdPatch$Json$Params {
  id: number;
      body?: {
'StudentName'?: string;
'StudentCode'?: string;
'File'?: Blob;
}
}

export function apiSubmissionsIdPatch$Json(http: HttpClient, rootUrl: string, params: ApiSubmissionsIdPatch$Json$Params, context?: HttpContext): Observable<StrictHttpResponse<SubmissionDto>> {
  const rb = new RequestBuilder(rootUrl, apiSubmissionsIdPatch$Json.PATH, 'patch');
  if (params) {
    rb.path('id', params.id, {});
    rb.body(params.body, 'multipart/form-data');
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'text/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<SubmissionDto>;
    })
  );
}

apiSubmissionsIdPatch$Json.PATH = '/api/submissions/{id}';